Move all toast-related source into a toast subpackage.
