+++ sotodlib/toast/pipeline_tools/observation.py
