+++ sotodlib/toast/time_constant.py
